pyproject.toml
[project]
name = "a2a-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.2.5",
    "agno[all]>=1.5.5",
    "asyncclick>=8.1.8",
    "click>=8.2.1",
    "fastapi>=0.115.12",
    "firecrawl>=2.7.1",
    "google-genai>=1.17.0",
    "googlesearch-python>=1.3.0",
    "httpx>=0.28.1",
    "jwt>=1.3.1",
    "langchain-cerebras>=0.5.0",
    "langchain-core>=0.3.62",
    "langchain-google-genai>=2.1.5",
    "langgraph>=0.4.7",
    "loguru>=0.7.3",
    "pydantic>=2.11.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "youtube-transcript-api>=1.0.3",
]

[tool.uv.workspace]
members = ["crewai_agen5ts", "agno_agents", "llama_index_agents", "pydantic_ai_agent"]
